Updated GitHub auth hook; Renamed Google auth hook
