+++ client/features/auth-pages/SignUpPage.tsx
@@ -3,8 +3,8 @@ import { Button, Box, Typography, Link, TextField } from '@mui/material';
 import { useNavigate, Link as RouterLink } from 'react-router-dom';
 import useDocumentTitle from '@hooks/useDocumentTitle';
 import styles from '@styles/layout/AuthStyles.module.css';
-import { useGoogleAuthMutation } from '@/hooks/googleAuthMutation';
-import { useGithubAuthMutation } from '@/hooks/useSignInWithGitHub';
+import { useGoogleAuthMutation } from '@hooks/useSignInWithGoogle';
+import { useGithubAuthMutation } from '@hooks/useSignInWithGitHub';
 import { useCreateUserWithEmailAndPasswordMutation } from '@/hooks/useCreateUserWithEmailAndPassword';
 import type { SubmitHandler } from 'react-hook-form';
 import { useForm } from 'react-hook-form';
